# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Platform Demo API — a multi-tenant Platform that hosts many independent
practitioners, each of whom connects their own OnFire Health partner account
via OAuth. The Platform lists each practitioner's rate cards, creates
invoices on their behalf, and receives invoice webhooks at one routed
receiver.

 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface SessionState {
  authenticated: boolean;
  /** @nullable */
  operator?: string | null;
}

export interface OperatorCredentials {
  /** @minLength 1 */
  password: string;
}

export interface PlatformConfig {
  configured: boolean;
  redirectUri: string;
  webhookUrl: string;
  scopes: string;
  /** @nullable */
  authorizeUrl?: string | null;
  missing: string[];
}

export interface PlatformSummary {
  totalConnections: number;
  activeConnections: number;
  totalInvoices: number;
  paidInvoices: number;
  openInvoices: number;
  webhookEvents: number;
}

export interface Connection {
  id: string;
  partnerPublicId: string;
  displayName: string;
  /** @nullable */
  orgId?: string | null;
  /** @nullable */
  scope?: string | null;
  /** active | revoked */
  status: string;
  connectedAt: string;
  /** @nullable */
  expiresAt?: string | null;
}

export interface AuthorizeInput {
  /** @minLength 1 */
  displayName: string;
}

export interface AuthorizeUrl {
  url: string;
}

export interface RateCard {
  refId: string;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  amount?: string | null;
  /** @nullable */
  currency?: string | null;
  active: boolean;
}

export interface Invoice {
  id: string;
  connectionId: string;
  externalInvoiceRef: string;
  /** @nullable */
  invoicePublicId?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  amount?: string | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  clientEmail?: string | null;
  /** @nullable */
  rateCardRefId?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface InvoiceInput {
  /** @minLength 1 */
  rateCardRefId: string;
  /** @minLength 1 */
  clientEmail: string;
  clientFirstName?: string;
  clientLastName?: string;
  clientPhone?: string;
  clientBillingAddress?: string;
  externalClientRef?: string;
}

export interface WebhookEvent {
  id: string;
  envelopeId: string;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  partnerPublicId?: string | null;
  /** @nullable */
  externalInvoiceRef?: string | null;
  /** @nullable */
  invoicePublicId?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  amount?: string | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  connectionId?: string | null;
  /** @nullable */
  connectionDisplayName?: string | null;
  routed: boolean;
  receivedAt: string;
}