# lib/api-zod/src/generated/types/invoice.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Platform Demo API — a multi-tenant Platform that hosts many independent
practitioners, each of whom connects their own OnFire Health partner account
via OAuth. The Platform lists each practitioner's rate cards, creates
invoices on their behalf, and receives invoice webhooks at one routed
receiver.

 * OpenAPI spec version: 0.1.0
 */

export interface Invoice {
  id: string;
  connectionId: string;
  externalInvoiceRef: string;
  /** @nullable */
  invoicePublicId?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  amount?: string | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  clientEmail?: string | null;
  /** @nullable */
  rateCardRefId?: string | null;
  createdAt: Date;
  updatedAt: Date;
}